backend/requirements.txt
mysql-connector-python~=9.2.0
python-dotenv~=1.1.0
faiss-cpu~=1.10.0
numpy~=2.0.2
protobuf~=5.29.4
streamlit~=1.44.1
bs4~=0.0.2
beautifulsoup4~=4.13.3
langchain~=0.3.23
flask==1.0.2